requirements.txt
streamlit==1.38.0
PyPDF2==3.0.1
langchain==0.0.235
openai==0.27.10
pinecone-client==2.2.1
